Revert "File upload"

+++ Models/Book.cs
@@ -11,8 +11,11 @@ namespace GooBitAPI.Models
 
         [BsonElement("Name")]
         public string BookName { get; set; } = null!;
-        public int Price { get; set; }
+
+        public decimal Price { get; set; }
+
         public string Category { get; set; } = null!;
+
         public string Author { get; set; } = null!;
     }
 }
